:recycle:refactor: 담당자 드롭다운 컴포넌트 분리
